Progress with enable decimal testing

+++ Core/DTOs/Transfer/TransferUpdateContentTargetDetailRequest.cs
@@ -5,8 +5,8 @@ namespace Core.DTOs.Transfer;
 public class TransferUpdateContentTargetDetailRequest {
     [Required]
     public Guid ID { get; set; }
-    
+
     public List<Guid>? RemoveRows { get; set; }
-    
-    public Dictionary<Guid, int>? QuantityChanges { get; set; }
+
+    public Dictionary<Guid, decimal>? QuantityChanges { get; set; }
 }
