+++ UnitTests/Integration/ExternalSystems/InventoryCounting/InventoryCountingDecreaseSystemBinTestHelpers/AddItems.cs
@@ -16,7 +16,7 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
     private readonly List<(int binEntry, string binCode, int quantity, UnitType unit)> binEntries = [];
 
     private Guid lastLineId = Guid.Empty;
-    private int  lastBinEntry;
+    private int lastBinEntry;
 
     public async Task<List<(int binEntry, string binCode, int quantity, UnitType unit)>> Execute() {
         await LoadBins();
@@ -29,7 +29,7 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
 
     private async Task LoadBins() {
         string connectionString = configuration.GetConnectionString("ExternalAdapterConnection") ?? throw new InvalidOperationException("ExternalAdapterConnection not found in configuration");
-        string query            = $"select top 4 \"AbsEntry\", \"BinCode\" from OBIN where \"WhsCode\" = '{testWarehouse}' and \"AbsEntry\" <> {testBinLocation} order by NEWID()";
+        string query = $"select top 4 \"AbsEntry\", \"BinCode\" from OBIN where \"WhsCode\" = '{testWarehouse}' and \"AbsEntry\" <> {testBinLocation} order by NEWID()";
         try {
             await using var connection = new SqlConnection(connectionString);
             await connection.OpenAsync();
@@ -64,16 +64,17 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
 
     private async Task AddItem() {
         foreach (var entry in binEntries) {
-            using var scope   = factory.Services.CreateScope();
-            var       service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsLineService>();
+            using var scope = factory.Services.CreateScope();
+            var service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsLineService>();
             var response = await service.AddItem(TestConstants.SessionInfo, new InventoryCountingAddItemRequest() {
-                BarCode  = testItem,
+                BarCode = testItem,
                 BinEntry = entry.binEntry,
-                ID       = id,
+                ID = id,
                 ItemCode = testItem,
                 Quantity = entry.quantity,
-                Unit     = entry.unit
+                Unit = entry.unit
             });
+
             Assert.That(response, Is.Not.Null);
             Assert.That(response.LineId, Is.Not.Null);
             Assert.That(response.Status, Is.EqualTo(ResponseStatus.Ok));
@@ -83,14 +84,15 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
     }
 
     private async Task ValidateCountingContent(bool afterUpdate = false) {
-        using var scope   = factory.Services.CreateScope();
-        var       service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsService>();
+        using var scope = factory.Services.CreateScope();
+        var service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsService>();
         var request = new InventoryCountingContentRequest {
             ID = id,
         };
-        var       responses        = await service.GetCountingContent(request);
+
+        var responses = await service.GetCountingContent(request);
         foreach (var row in responses) {
-            var entry         = binEntries.First(b => b.binEntry == row.BinEntry);
+            var entry = binEntries.First(b => b.binEntry == row.BinEntry);
             // Assert.That(row.BinCode, Is.EqualTo(entry.binCode)); TODO: Bin Code is not loaded yet
             Assert.That(row.ItemCode, Is.EqualTo(testItem));
             Assert.That(row.ItemName, Is.EqualTo($"Test Item {testItem}"));
@@ -98,15 +100,17 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
             Assert.That(row.NumInBuy, Is.EqualTo(12));
             Assert.That(row.PurPackMsr, Is.EqualTo("Box"));
             Assert.That(row.PurPackUn, Is.EqualTo(4));
-            
-            int entryQuantity = entry.quantity;
+
+            decimal entryQuantity = entry.quantity;
             if (afterUpdate && entry.binEntry == lastBinEntry)
                 entryQuantity = 1;
+
             if (entry.unit != UnitType.Unit)
                 entryQuantity *= row.NumInBuy;
+
             if (entry.unit == UnitType.Pack)
                 entryQuantity *= row.PurPackUn;
-            
+
             Assert.That(row.CountedQuantity, Is.EqualTo(entryQuantity));
             Assert.That(row.SystemQuantity, Is.Zero);
             Assert.That(row.Variance, Is.EqualTo(entryQuantity));
@@ -114,14 +118,15 @@ public class AddItems(Guid id, string testItem, string testWarehouse, WebApplica
     }
 
     private async Task UpdateItem() {
-        using var scope   = factory.Services.CreateScope();
-        var       service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsLineService>();
+        using var scope = factory.Services.CreateScope();
+        var service = scope.ServiceProvider.GetRequiredService<IInventoryCountingsLineService>();
         var request = new InventoryCountingUpdateLineRequest {
-            Id       = id,
-            LineId   = lastLineId,
+            Id = id,
+            LineId = lastLineId,
             Quantity = 1,
-            Comment  = "Changed Quantity",
+            Comment = "Changed Quantity",
         };
+
         var response = await service.UpdateLine(TestConstants.SessionInfo, request);
         Assert.That(response, Is.Not.Null);
         Assert.That(response.ReturnValue, Is.EqualTo(UpdateLineReturnValue.Ok));
